wifi: ath11k: Fix 2 second busywait during stats request
This diff fixes a 2-second busy loop inside
the ath11k_debugfs_fw_stats_request function. The code in question was
added during the initial import of the ath11 driver. While the ath11 code
is based on ath10, the latter doesn't have this bug. This diffs changes
the wait loop to be similar to the ath10 driver. It also adds an explicit
delay before retrying. An explicit delay avoids spamming FW with stats
requests.